# openapi.yaml
openapi: "3.1.0"
info:
  title: Agent Kanban REST API
  description: REST API for managing Kanban projects, tickets, columns, workflows, roles, agent communications (conversations), and global settings. Each MCP server process is associated with a role that determines permissions.
  version: "1.0.0"
  contact:
    name: Agent Kanban Team
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
paths:
...
  /projects/{slug}/columns:
    get:
      summary: List columns for a project
      tags:
        - Columns
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of columns
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ColumnListResponse'
    post:
      summary: Add column to a project
      tags:
        - Columns
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateColumnRequest'
      responses:
        '201':
          description: Column created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ColumnResponse'
  /projects/{slug}/columns/{columnId}:
    patch:
      summary: Update a column
      tags:
        - Columns
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: columnId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateColumnRequest'
      responses:
        '200':
          description: Column updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ColumnResponse'
    delete:
      summary: Delete a column (fails if non-empty or default)
      tags:
        - Columns
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: columnId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Column deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '409':
          $ref: '#/components/responses/ConflictError'

  # ==========================================================================
  # WORKFLOW (project-level)
  # ==========================================================================
...
  /projects/{slug}/roles:
    get:
      summary: Get all roles with access levels for a project
      tags:
        - Roles
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of roles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleListResponse'
  /projects/{slug}/roles/{roleId}:
    patch:
      summary: Update a role (name/description)
      tags:
        - Roles
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: roleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
      responses:
        '200':
          description: Role updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'

  # ==========================================================================
  # ROLES-COLUMNS (project-level)
  # ==========================================================================
  /projects/{slug}/roles-columns:
    get:
      summary: Get combined project and global roles_columns mappings
      tags:
        - Roles
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of roles_columns mappings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolesColumnsListResponse'
    post:
      summary: Create a project-specific roles_columns override
      tags:
        - Roles
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRolesColumnRequest'
      responses:
        '201':
          description: Roles-column mapping created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolesColumnResponse'
  /projects/{slug}/roles-columns/{roleId}:
    patch:
      summary: Update a project-specific roles_columns override
      tags:
        - Roles
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: roleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRolesColumnRequest'
      responses:
        '200':
          description: Roles-column mapping updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolesColumnResponse'
    delete:
      summary: Delete a project-specific roles_columns override
      tags:
        - Roles
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: roleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Roles-column mapping deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean

  # ==========================================================================
  # TICKETS
  # ==========================================================================
...
components:
  schemas:
...
    StandardSuccessResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          nullable: true
        error:
          type: string
          nullable: true
          example: null
    StandardErrorResponse:
      type: object
      required:
        - success
        - error
        - code
      properties:
        success:
          type: boolean
          example: false
        data:
          nullable: true
        error:
          type: string
          example: "Resource not found"
        code:
          type: string
          example: "NOT_FOUND"

    # ----- Project -----
...
    Column:
      type: object
      properties:
        id:
          type: integer
          example: 8
        project_id:
          type: integer
          example: 1
        slug:
          type: string
          description: URL-friendly column identifier
          example: "todo"
        name:
          type: string
          description: Human-readable column name
          example: "To Do"
        order:
          type: integer
          minimum: 0
          example: 0
        is_global:
          type: boolean
          example: false
        is_default:
          type: boolean
          example: true
    CreateColumnRequest:
      type: object
      required:
        - slug
        - name
        - order
      properties:
        slug:
          type: string
          minLength: 1
          maxLength: 255
        name:
          type: string
          minLength: 1
          maxLength: 255
        order:
          type: integer
          minimum: 0
        is_default:
          type: boolean
          default: false
    UpdateColumnRequest:
      type: object
      properties:
        slug:
          type: string
          minLength: 1
          maxLength: 255
        name:
          type: string
          minLength: 1
          maxLength: 255
        order:
          type: integer
          minimum: 0

    # ----- Global Column -----
...
    Role:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Human User"
        description:
          type: string
          example: "Full permissions for human users"
        access_level:
          type: string
          enum:
            - full
            - admin
            - write
            - read
          example: "full"
...
    UpdateRoleRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 2000

    # ----- Global Role -----
...
    RolesColumn:
      type: object
      properties:
        id:
          type: integer
        role_id:
          type: integer
        project_id:
          type: integer
          nullable: true
        column_id:
          type: integer
          nullable: true
        is_default:
          type: integer
          example: 0
    CreateRolesColumnRequest:
      type: object
      required:
        - role_id
        - column_id
      properties:
        role_id:
          type: integer
        column_id:
          type: integer
        is_default:
          type: integer
          default: 0
    UpdateRolesColumnRequest:
      type: object
      properties:
        column_id:
          type: integer
        is_default:
          type: integer
          default: 0

    # ----- Global Roles-Columns -----
...
    ColumnListResponse:
      allOf:
        - $ref: '#/components/schemas/StandardSuccessResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Column'
    ColumnResponse:
      allOf:
        - $ref: '#/components/schemas/StandardSuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Column'

    # Global Column responses
...
    RoleListResponse:
      allOf:
        - $ref: '#/components/schemas/StandardSuccessResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Role'
    RoleResponse:
      allOf:
        - $ref: '#/components/schemas/StandardSuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Role'

    # Global Role responses
...
    RolesColumnsListResponse:
      allOf:
        - $ref: '#/components/schemas/StandardSuccessResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/RolesColumn'
    RolesColumnResponse:
      allOf:
        - $ref: '#/components/schemas/StandardSuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/RolesColumn'
...
    SuccessResponse:
      allOf:
        - $ref: '#/components/schemas/StandardSuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                deleted:
                  type: boolean
  responses:
...
    ConflictError:
      description: Conflict (e.g., column has tickets, circular dependency)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
...